+++ src/dipdup/indexes/evm_subsquid_events/index.py
@@ -76,7 +76,7 @@ class SubsquidEventsIndex(
         return self._topics
 
     async def _process_queue(self) -> None:
-        logs_by_level = defaultdict(list)
+        logs_by_level: dict[int, list[EvmNodeLogData]] = defaultdict(list)
 
         # NOTE: Drain queue and group messages by level.
         while True:
@@ -152,14 +152,20 @@ class SubsquidEventsIndex(
                 typename = handler.contract.module_name
                 topics.add(self.topics[typename][handler.name])
 
-            # NOTE: Requesting logs by batches of NODE_BATCH_SIZE.
+            # NOTE: Requesting logs by batches of datasource_config.http.batch_size or NODE_BATCH_SIZE.
+            evm_node: EvmNodeDatasource = self.random_node
+            if evm_node._http_config:
+                batch_size = evm_node._http_config.batch_size
+            else:
+                batch_size = NODE_BATCH_SIZE
+
             batch_first_level = first_level
             while batch_first_level <= sync_level:
                 # NOTE: We need block timestamps for each level, so fetch them separately and match with logs.
                 timestamps: dict[int, int] = {}
                 tasks: deque[asyncio.Task[Any]] = deque()
 
-                batch_last_level = min(batch_first_level + NODE_BATCH_SIZE, sync_level)
+                batch_last_level = min(batch_first_level + batch_size, sync_level)
                 level_logs_task = asyncio.create_task(
                     self.random_node.get_logs(
                         {
@@ -168,13 +174,14 @@ class SubsquidEventsIndex(
                         }
                     )
                 )
-                tasks.append(level_logs_task)
 
                 async def _fetch_timestamp(level: int, timestamps: dict[int, int]) -> None:
-                    block = await self.random_node.get_block_by_level(level)
+                    block = await self.random_node.get_block_by_level(level, full_transactions_data=False)
                     timestamps[level] = int(block['timestamp'], 16)
 
-                for level in range(batch_first_level, batch_last_level + 1):
+                level = batch_last_level
+                level_logs = await level_logs_task
+                for level in [int(log['blockNumber'], 16) for log in level_logs]:
                     tasks.append(
                         asyncio.create_task(
                             _fetch_timestamp(level, timestamps),
@@ -184,7 +191,6 @@ class SubsquidEventsIndex(
 
                 await asyncio.gather(*tasks)
 
-                level_logs = await level_logs_task
                 parsed_level_logs = tuple(
                     EvmNodeLogData.from_json(
                         log,
@@ -197,6 +203,7 @@ class SubsquidEventsIndex(
                 Metrics.set_sqd_processor_last_block(level)
 
                 batch_first_level = batch_last_level + 1
+                await asyncio.sleep(evm_node._http_config.ratelimit_sleep)
         else:
             sync_level = min(sync_level, subsquid_sync_level)
             fetcher = self._create_fetcher(first_level, sync_level)