+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 
 ## [Unreleased]
 
+### Performance
+
+- evm.subsquid.events: EvmNodeDatasource._synchronize() optimisations
+
 ### Fixed
 
 - ci: Fixed falling `test_metadata_networks()`